subtitle_generator/requirements.txt
asgiref==3.7.2
asttokens==2.4.0
backcall==0.2.0
beautifulsoup4==4.12.2
cachetools==5.3.1
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.2.0
colorama==0.4.6
decorator==4.4.2
Django==4.2.5
djangorestframework==3.14.0
exceptiongroup==1.1.3
executing==1.2.0
filelock==3.12.4
google==3.0.0
google-api-core==2.11.1
google-auth==2.23.0
google-cloud-core==2.3.3
google-cloud-storage==2.11.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
imageio==2.31.3
imageio-ffmpeg==0.4.9
ipython==8.15.0
jedi==0.19.0
Jinja2==3.1.2
llvmlite==0.41.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
more-itertools==10.1.0
moviepy==2.0.0.dev2
mpmath==1.3.0
networkx==3.1
numba==0.58.0
numpy==1.25.2
openai-whisper @ git+https://github.com/openai/whisper.git@0a60fcaa9b86748389a656aa013c416030287d47
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.1
proglog==0.1.10
prompt-toolkit==3.0.39
protobuf==4.24.3
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.16.1
pytz==2023.3.post1
regex==2023.8.8
requests==2.31.0
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.2
sympy==1.12
tiktoken==0.3.3
torch==2.0.1
tqdm==4.66.1
traitlets==5.10.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.16
wcwidth==0.2.6
